Revert "Dep/General: fix some strict c++11 warnings/errors"
This reverts commit 7a41bbf75cfdd3cae0889d86888b80f82be35c67.

+++ src/server/authserver/Server/RealmSocket.cpp
@@ -96,7 +96,7 @@ int RealmSocket::open(void * arg)
     return 0;
 }
 
-int RealmSocket::close(u_long)
+int RealmSocket::close(int)
 {
     shutdown();
